extracted logic to load the icons
