+++ AnnoDesigner/AnnoDesigner.csproj
@@ -52,6 +52,7 @@
       <DependentUpon>About.xaml</DependentUpon>
     </Compile>
     <Compile Include="Commons.cs" />
+    <Compile Include="PresetsLoader\IconLoader.cs" />
     <Compile Include="PresetsLoader\IconMappingPresetsLoader.cs" />
     <Compile Include="model\ICommons.cs" />
     <Compile Include="model\IUpdateHelper.cs" />
